refactor: moved stac search logic to StacCollection class
